Moved login and register to auth
